requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
mimesis==13.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
jinja2==3.1.2